chore: MOJI 프로젝트 폴더 아키텍처 구성 및 테스트 코드 작성
