+++ build.gradle
@@ -15,8 +15,7 @@ java {
 }
 
 configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
+    compileOnly {        extendsFrom annotationProcessor
     }
 }
