--- src/main/java/com/spring/moji/domain/test/dto/request/TestRequestDTO.java
@@ -1,11 +1,0 @@
-package com.spring.moji.domain.test.dto.request;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class TestRequestDTO {
-	private String id;
-	private String password;
-}
