--- src/main/java/com/spring/moji/domain/test/dto/response/TestResponseDTO.java
@@ -1,10 +1,0 @@
-package com.spring.moji.domain.test.dto.response;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class TestResponseDTO {
-	private String id;
-}
