+++ src/main/java/com/spring/moji/dto/request/TestRequestDTO.java
@@ -1,0 +1,11 @@
+package com.spring.moji.dto.request;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class TestRequestDTO {
+	private String id;
+	private String password;
+}
