+++ src/main/java/com/spring/moji/dto/response/TestResponseDTO.java
@@ -1,0 +1,10 @@
+package com.spring.moji.dto.response;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class TestResponseDTO {
+	private String id;
+}
