+++ src/main/resources/static/css/common/layout.css
@@ -1,0 +1,224 @@
+/* 공통 스타일 */
+body {
+    font-weight: normal;
+    font-style: normal;
+    display: flex;
+    min-height: 100vh;
+    background-color: #fffae7;
+}
+
+/* 헤더 스타일 */
+header {
+    background-color: #ffffff;
+    padding: 20px;
+    display: flex;
+    justify-content: space-between; /* h1이 menu-bar 옆에 위치하도록 설정 */
+    align-items: center; /* 수직 가운데 정렬 */
+    border-bottom: 1px solid #ddd;
+}
+
+header h1 {
+    margin: 0;
+    color: #2d2d2b;
+}
+
+/* 메뉴바 스타일 */
+.menu-bar {
+    cursor: pointer;
+    margin-right: 10px;
+    flex-shrink: 0; /* 이미지 크기가 줄어들지 않도록 방지 */
+}
+
+/* 사이드바 스타일 */
+.sidebar {
+    width: 250px;
+    transition: width 0.3s ease;
+    background-color: #ffeeb3;
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    border-right: 1px solid #ddd;
+    position: fixed;
+    top: 0;
+    left: 0;
+    height: 100vh;
+}
+
+.sidebar.collapsed {
+    width: 100px; /* 아이콘만 보일 때의 너비 */
+    padding: 50px;
+}
+
+.sidebar.collapsed .profile {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.sidebar.collapsed .profile img {
+    margin-bottom: 10px;
+}
+
+.sidebar.collapsed nav ul li a {
+    justify-content: center; /* 아이콘을 가운데 정렬 */
+}
+
+.sidebar.collapsed nav ul li a img {
+    margin: 0;
+}
+
+.sidebar.collapsed nav ul li a span {
+    display: none; /* 사이드바 축소 시 텍스트 숨기기 */
+}
+
+/* 프로필 스타일 */
+.profile {
+    text-align: center;
+    margin-bottom: 40px;
+}
+
+.profile img {
+    width: 70px;
+    height: 70px;
+    border-radius: 50%;
+    margin-bottom: 10px;
+}
+
+.profile h3 {
+    margin-bottom: 5px;
+    color: #2d2d2b;
+}
+
+.profile p {
+    font-size: 14px;
+    color: #888;
+}
+
+/* 네비게이션 스타일 */
+nav ul {
+    list-style: none;
+    width: 100%;
+}
+
+nav ul li {
+    width: 100%;
+    margin-bottom: 20px;
+}
+
+nav ul li a {
+    text-decoration: none;
+    color: #2d2d2b;
+    font-weight: bold;
+    display: flex;
+    align-items: center;
+    padding: 10px;
+    text-align: left;
+    transition: background 0.3s;
+}
+
+nav ul li a img {
+    margin-right: 10px;
+}
+
+nav ul li a:hover {
+    background-color: #d4c0e6;
+    border-radius: 8px;
+}
+
+nav ul li a span {
+    transition: opacity 0.3s;
+}
+
+.sidebar.collapsed nav ul li a span {
+    opacity: 0; /* 텍스트 숨기기 */
+}
+
+/* 메인 콘텐츠 스타일 */
+.main-content {
+    flex-grow: 1;
+    background-color: #f3f3f3;
+    padding: 20px;
+    transition: margin-left 0.3s ease;
+    margin-left: 250px; /* 사이드바 크기만큼 이동 */
+}
+
+.main-content.shifted {
+    margin-left: 100px; /* 사이드바가 축소되었을 때 */
+}
+
+/* 사용자 정보 및 알림 스타일 */
+.user-info {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+}
+
+.user-info span {
+    margin-right: 20px;
+    font-weight: bold;
+    color: #2d2d2b;
+}
+
+.notification {
+    cursor: pointer;
+    font-size: 24px;
+    color: #888;
+}
+
+/* 드롭다운 스타일 */
+.custom-dropdown {
+    margin-top: -10px;
+    background-color: transparent; /* 배경색 제거 */
+    border: none; /* 테두리 제거 */
+    padding: 0; /* 여백 제거 */
+    box-shadow: none; /* 그림자 제거 */
+}
+
+.custom-dropdown:focus {
+    box-shadow: none; /* 포커스 시 그림자 제거 */
+}
+
+/* 로그인/회원가입 링크 스타일 */
+.auth-link {
+    text-decoration: none; /* 밑줄 제거 */
+    color: #333; /* 링크 색상 설정 */
+    font-weight: bold;
+    margin: 0 5px;
+}
+
+.auth-link:hover {
+    color: #d4c0e6; /* 마우스를 올렸을 때 색상 변경 */
+    text-decoration: underline; /* 마우스를 올렸을 때 밑줄 표시 */
+}
+
+/* 푸터 스타일 */
+.footer {
+    bottom: 0;
+    left: 0;
+    /*width: calc(100% - 250px); !* 기본 사이드바 너비 제외한 나머지 너비 *!*/
+    width: 100%;
+    /*margin-right: 250px; !* 사이드바의 기본 너비만큼 왼쪽에서 간격 *!*/
+    text-align: center;
+    padding: 10px 0;
+    border-top: 1px solid #ddd;
+    transition: margin-left 0.3s ease, width 0.3s ease;
+}
+
+.content {
+    margin-top: 10px;
+    padding: 10px;
+    background-color: #ffffff;
+    width: 100%;
+    min-height: calc(100vh - 25%); /* 전체 높이에서 헤더와 푸터를 제외한 높이 */
+    overflow-y: auto; /* 콘텐츠가 길어지면 스크롤 */
+}
+
+.content > div {
+    min-height: 100%;
+    width: 100%; /* 부모인 content의 100% 너비 */
+    max-width: 1200px; /* 최대 너비 설정 */
+    padding: 20px;
+    box-sizing: border-box;
+}
