+++ src/main/resources/static/css/content/user/profile-modal-style.css
@@ -1,0 +1,57 @@
+.modal {
+  display: none; /* Hidden by default */
+  position: fixed; /* Stay in place */
+  z-index: 1; /* Sit on top */
+  left: 0;
+  top: 0;
+  width: 100%; /* Full width */
+  height: 100%; /* Full height */
+  overflow: auto; /* Enable scroll if needed */
+  background-color: rgb(0, 0, 0); /* Fallback color */
+  background-color: rgba(0, 0, 0, 0.4); /* Black w/ opacity */
+  animation-name: fade-in;
+  animation-duration: 0.4s;
+}
+
+.modal .btn-success {
+  background-color: #ffe898;
+  width: 156px;
+  color: black;
+  border: 0px;
+  font-size: 1rem;
+  text-align: center;
+}
+
+.modal .btn-success:hover {
+  background-color: #fad13e;
+  box-shadow: 0 0 10px #ffc80073 !important;
+}
+
+.modal .btn-success:active,
+.modal .btn-success:focus {
+  background-color: #fad13e;
+}
+
+.modal-content {
+  position: absolute;
+  bottom: 50%;
+  left: 50%;
+  transform: translate(-50%, 50%);
+  background-color: #fefefe;
+  width: 700px;
+  animation-name: slide-in;
+  animation-duration: 0.4s;
+}
+
+.modal-content * {
+  margin-bottom: 5px;
+}
+
+@keyframes fade-in {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
