+++ src/main/resources/static/css/content/user/profile-style.css
@@ -1,0 +1,114 @@
+.hidden {
+  display: none !important;
+}
+
+.main-panel {
+  background-color: #ffeeb1;
+  width: 100%;
+  height: 100%;
+}
+
+.main-panel .content-wrapper {
+  padding: 2.75rem 30%;
+}
+
+.main-panel .content-wrapper .profile-card {
+  width: 100%;
+  min-width: 500px;
+  padding: 40px;
+  margin: 0 auto;
+  background-color: #fff;
+  box-shadow: 10px 10px 30px rgba(0, 0, 0, 0.1);
+  border-radius: 10px;
+  align-items: center;
+  display: block;
+}
+
+.profile-content .triple-pfp-container {
+  display: flex;
+  align-items: baseline;
+  justify-content: center;
+  margin-bottom: 15px;
+  gap: 10px;
+}
+
+.profile-content .pfp-background {
+  overflow: visible;
+  width: 110px;
+  height: 110px;
+  border-radius: 50%;
+  background-color: #ffe898;
+  display: inline-block;
+  position: relative;
+}
+
+.profile-content .individual {
+  width: 60px;
+  height: 60px;
+}
+
+.profile-content .pfp-background .bi {
+  position: absolute;
+  top: 69%;
+  left: 70%;
+  width: 30px;
+  height: 30px;
+  background-color: rgb(241, 241, 241);
+  border-radius: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 1.2rem;
+}
+
+.profile-content .pfp-background .pfp-container {
+  overflow: hidden;
+  width: 90%;
+  height: 90%;
+  border-radius: 50%;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  cursor: pointer;
+}
+
+.profile-content .pfp-background .pfp-container .pfp {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.profile-card .btn-success {
+  background-color: #ffe898;
+  width: 156px;
+  color: black;
+  border: 0px;
+  font-size: 1rem;
+  text-align: center;
+}
+
+.profile-card .btn-success {
+  display: inline-block;
+}
+
+.profile-card .btn-success img {
+  padding-top: 10px;
+  width: 30px;
+  display: block;
+  margin: 0 auto;
+}
+
+.profile-card .btn-success:hover {
+  background-color: #fad13e;
+  box-shadow: 0 0 10px #ffc80073 !important;
+}
+.profile-card .btn-success:active,
+.profile-card .btn-success:focus {
+  background-color: #fad13e;
+}
+
+.profile-card .btn-container {
+  display: flex;
+  gap: 20px;
+}
