--- src/main/webapp/WEB-INF/component/test-header.jsp
@@ -1,23 +1,0 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="stylesheet" type="text/css"  href="${pageContext.request.contextPath}/css/component/testHeader.css">
-</head>
-<body>
-    <header>
-        <div class="header">
-            <nav>
-                <ul>
-                    <li><a href="/home">Home</a></li>
-                    <li><a href="/profile">Profile</a></li>
-                    <li><a href="/settings">Settings</a></li>
-                    <li><a href="/logout">Logout</a></li>
-                </ul>
-            </nav>
-        </div>
-        <link rel="stylesheet" href="/resources/css/header.css">
-    </header>
-</body>
-</html>
