+++ src/main/webapp/WEB-INF/jsp/page/user/profile-page.jsp
@@ -1,0 +1,2 @@
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<jsp:include page="/WEB-INF/jsp/layout/layout.jsp" />
