--- src/main/webapp/WEB-INF/page/test/test-home.jsp
@@ -1,11 +1,0 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
-<html>
-<head>
-    <title>Title</title>
-</head>
-<body>
-    <%@ include file="/WEB-INF/component/test-header.jsp" %>
-    <h1>This is JSP / SPRING BOOT TestPage!</h1>
-</body>
-</html>
